// sdk/packages/realm/src/models/WorldEventListDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { WorldEventDetailDto } from './WorldEventDetailDto';
export type WorldEventListDto = {
    editorSnapshotVersion?: string;
    eventGraphSummary?: Record<string, any>;
    items: Array<WorldEventDetailDto>;
    worldId: string;
};