// sdk/packages/realm/src/models/WorldMutationListDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { WorldMutationSummaryDto } from './WorldMutationSummaryDto';
export type WorldMutationListDto = {
    items: Array<WorldMutationSummaryDto>;
    worldId: string;
};